Reorganize dotfiles into directory structure for use with stow
The bare repo solution I was using previously is just too annoying
for me.  I don't like to use git in the terminal because I like to be
able to interactively stage hunks, see the branch structure, and so on.
Using a bare repo separates the worktree from the content of the .git
folder and requires that all git commands have --git-dir="$HOME/.cfg"
and --work-tree="$HOME" appended to every command.  It therefore does
not work with any git GUIs.

Stow can be used to automatically create symlinks for each bit of
software, including in nested directories, and keeps all dotfiles in a
normal (working) git repo, which is better for my workflow.